Handle no dhcpd settings when upgrading
I was testing 2.1.4-release on a new full install in a VM with only WAN defined. During the first boot after install this warning is emitted:
Warning: Invalid argument supplied for foreach() in /etc/inc/upgrade_config.inc on line 3153
It does not cause any issue, but might as well properly handle the case when there are no dhcpd settings.