Handle no dhcpd settings when upgrading
This minor fix was in master but not 2.1 branch. I noticed the warning message when doing a fresh install/test of 2.1.4-release. It prevents the warning message:
Warning: Invalid argument supplied for foreach() in /etc/inc/upgrade_config.inc on line 3153
Might as well fix it it 2.1-branch also, in case there is another 2.1.n release.