Standardise size of Duplicate Slice button
The Duplicate Slice button currently is displayed in smaller text and in a row of its own, separate from the row above that has the rest of the "Duplicate bootup slice" text and slice selection.
This change puts the button in the same row as the slice selection and text, and makes the button text be the same size as the text in other buttons on this page.
Did this first on a 2.1.5 system, so have submitted this against 2.1 branch. I will make another pull request against master also, because a later xhtml compliance change in master will conflict.