Manage dhcpleaseinlocaltime consistently
dhcpleaseinlocaltime is actually a global setting, but the setting is stored per-DHCP-enabled-interface.
The display code in status_dhcp_leases already sorts this out - if any interface has the setting enabled then the displayed lease times are adjusted to local time.
This fixes up the input of the setting. If dhcpleaseinlocaltime is checked on any interface, then it will now be shown checked whichever interface tab the user happens to have open.
Whatever the user does to the checkbox, the setting is saved to all interfaces in the config - this is particularly neccesary when unchecking the box. The setting must be cleared in the config for all interfaces.
The way this change works, there is no need to convert existing configs. The code correctly works with both old and new configs.
This came up again in forum: https://forum.pfsense.org/index.php?topic=82881.msg0
It is an annoying "feature" so I think it is worth clearing it up on the UI.