Add option to use root FS for transitory XML files
On low memory systems this can be necessary if there are a few interfaces and thus a bit of RRD data. The XML files take up more space than the RRD. The space is only needed temporarily until the XML files are all packed into a tar.gz
This allows RRD data to save successfully on a low-memory where the limit of size of /var memory disk has been reached.