Display installed pkg version even if pkg server not available
Forum: https://forum.pfsense.org/index.php?topic=84820.0
It seems reasonable to me to display the currently installed package version numbers, along with text like "Latest: N/A". and the Version box being red-filled when someone goes to System->Packages and displays the installed packages but the package server is not reachable for some reason.
This code does it.
What do you think?