Explicitly pass gateway status and color
This code was passing a whole load of html to attempt to set various properties of the Gateways Widget Status column so that it would display the new status (Online, Offline...) and in the appropriate color.
Now that I am using class=listr, the css fro listr already specifies white. Setting bgcolor in HTML does not override that.
Pass both the status text and color as ordinary text, separated by "/". The consumer of this output can do their thing with jQuery to set the color as well as text of the element.