Standardise css list padding
On Status->Services, the Services Status widget when a service status is down, the red "x" button does not line up with the green "running" button in the column.
This is because listbg has:
padding-right: 16px;
But the normal listr has:
padding-right: 6px;
It seems really odd that listbg as 16 pixels right padding. I expected it to have all the same padding as listr, just with a different color.
listn and listbggrey have the same "feature" of extra padding-right.
listn does not seem to be used anywhere anyway - might as well "fix" it in case something uses it some time.
listbggrey seems to only be used by pkg_mgr_installed.php in the case when an installed package version is in advance of the online version - quite an unusual thing! Changing padding-right from 16px back to 6px is fine for that.
Then the "feature" is lso in listhdr and listhdrr - lots of examples like firewall_aliases.php have the column headings as listhdr or listhdrr, so in theory if the column is not so wide already, it will put 16px on the right after the column heading text. But the actual rows of data underneath use listr and listbg, which only specify 6px right padding. In the end it usually does not make any difference - there is space already and the browser lines it all up anyway. But it seems to me an odd difference in the css.
listtopic also has this 16px right padding thing - but it seems to be used just for text display that is not in columns, so I did not mess with that.
Changing this css fixes the lineup of the red "x" in the services status column and does not seem to break anything else. In theory it could make other things better also.
Someone who knows how the css was intended to be used could comment on this! If I have got the right idea, and these "16px" things were just a mistake from the past, then it should be checked in the css for all themes and corrected appropriately... 
I am happy to do that, once someone confirms it is the thing to do.