Update interfaces.php
Removed exclusion of 31 from subnet list boxes (lines 1771/1775 and 2007/2011). Starting in FreeBSD 9.x and above, the operating system supports /31 (255.255.255.254) subnets. This allows the use of point-to-point addresses for interfaces. Some ISP's (Cox Communications) are implementing IP-VPN services that use these 31-bit subnet masks. Customers are provided a point-to-point IP and a Gateway. There is no need for the 2 extra IP addresses that a /30 will give you, so a /31 is used instead.