OpenVPN client can have only 1 of username and/or password
In some situations it is valid to just have a Username or a Password, both are not always required. In that situation it does not matter if it is called "Username" or "Password" - it is just a (hopefully hard-to-guess) string used for authentication.
The actual validation code already allows this, so just clarify it in the descriptive text.