Remove "link", "vlink" and "alink" from BODY tag
This is a positional change to remove "link", "vlink" and "alink" from
the BODY tag, the following themes do not have Anchor tags defined in
CSS, so they use the colours defined in the BODY tag, which is blue
(hex: #0000CC).

_corporate
code-red
metallic
nervecenter
pfsense
pfsense-dropdown

Update "all.css" in the above themes with Anchor tags defined, so that
the colour definition can be removed from the BODY tag.