Interfaces widget remove blank line if no IPv4 address
If the interface had an IPv6 address but no IPv4 address, there was a blank line where the IPv4 address would have been. There is no need for that, and one day IPv4 will be old legacy and systems will routinely have no IPv4 addresses at all - they will all be IPv6. Might as well make that look ordinary on the display now.
The br goes in the div so we can put it in and out from the AJAX also.