Handle firewall log widget display formats
the same way for the initial display and for updated rows done by Java Script. Now we receive the source IP and port, destination IP and port, all in separate fields so they can be put together in whatever combination for display.
IPv6 displayed addresses are shown inside "[ ]" so that any following port has the standard syntax like "[a:b::c:d]:123" - this makes it obvious that the last numbers are a port number, and not part of the IPv6 address.
The "title" has IP+Port - that is displayed when hovering over the box in general.
The href to diag_dns.php has hover text "Reverse Resolve with DNS" and the "?host=" sends just the IP address (without IPv6 square brackets).
The text displayed in the link is the IP address (with square brackets if an IPv6 address).
For the destination column, if there is a destinaion port, it is displayed in ordinary text ":port" after the IP address.

The blank not-displayed row at the end of the table is removed - this fixes the problem with counting the rows of the table where rows would disappear at each update.