Adding boot delay to alleviate overwritten boot loader configs
Auto-upgrades on ALIX hardware for nanoBSD image results in an overwritten /boot/loader.conf. Upon rebooting, the image will not properly boot until the kernel boot delay is manually done via console and webConfig. Hopefully this will alleviate this issue, so I can upgrade all my pfSense nodes without having to travel to remote locations. Please let me know if this is the improper file to modify, or if this is not a feasible solution: lrallen03@gmail.com